Merge pull request #18 from MQuang01/hiep-it/api-user

+++ src/main/java/com/cghue/projecthousemaidwebapp/domain/OrderDetail.java
@@ -27,7 +27,7 @@ public class OrderDetail {
 
     private Integer quantity;
 
-    private Double houseSize;
+    private Integer houseSize;
 
     public OrderDetailResDto toResDto() {
         return new OrderDetailResDto(id, job.toResDto(), price, quantity, houseSize);
